Use the appropriate data type in the signature.

+++ source/Color.cpp
@@ -81,7 +81,7 @@ Color Color::Additive(float alpha) const
 	return result;
 }
 
-Color Color::Combine(double a1, Color c1, double a2, Color c2)
+Color Color::Combine(float a1, Color c1, float a2, Color c2)
 {
 	return Color(
 			a1 * c1.color[0] + a2 * c2.color[0],
